Upgrade and refactor
