+++ OsteklokkenServer/OsteklokkenServer/OsteSession.cs
@@ -1,6 +1,8 @@
-﻿namespace OsteklokkenServer
+﻿using Red.CookieSessions;
+
+namespace OsteklokkenServer
 {
-    public class OsteSession
+    public class OsteSession : CookieSessionBase
     {
         public string Username { get; set; }
         public string Name { get; set; }
